+++ src/app/Scenes/MyCollection/Screens/Artwork/MyCollectionArtworkAbout.tsx
@@ -6,7 +6,7 @@ import { Flex, Spacer } from "palette/elements"
 import React from "react"
 import { useFragment } from "react-relay"
 import { graphql } from "relay-runtime"
-import { showSubmitToSell } from "../../utils/checkShowSubmitToSell"
+import { showSubmitToSell } from "../../utils/checkArtworkDetails"
 import { MyCollectionArtworkAboutWork } from "./Components/ArtworkAbout/MyCollectionArtworkAboutWork"
 import { MyCollectionArtworkArticles } from "./Components/ArtworkAbout/MyCollectionArtworkArticles"
 import { MyCollectionArtworkPurchaseDetails } from "./Components/ArtworkAbout/MyCollectionArtworkPurchaseDetails"
