+++ src/app/Scenes/MyCollection/utils/checkArtworkDetails.tsx
@@ -3,20 +3,25 @@ import { MyCollectionArtworkInsights_artwork } from "__generated__/MyCollectionA
 import { OldMyCollectionArtworkQueryResponse } from "__generated__/OldMyCollectionArtworkQuery.graphql"
 import { STATUSES } from "../Screens/Artwork/Components/MyCollectionArtworkSubmissionStatus"
 
+export const isPOneArtist = (
+  artwork:
+    | MyCollectionArtworkAbout_artwork
+    | MyCollectionArtworkInsights_artwork
+    | NonNullable<OldMyCollectionArtworkQueryResponse["artwork"]>
+) =>
+  !!artwork.artists?.find((artist) => Boolean(artist?.targetSupply?.isTargetSupply)) ??
+  !!artwork.artist?.targetSupply?.isTargetSupply ??
+  false
+
 export const showSubmitToSell = (
   artwork:
     | MyCollectionArtworkAbout_artwork
     | MyCollectionArtworkInsights_artwork
     | NonNullable<OldMyCollectionArtworkQueryResponse["artwork"]>
 ) => {
-  const isPOneArtist =
-    !!artwork.artists?.find((artist) => Boolean(artist?.targetSupply?.isTargetSupply)) ??
-    !!artwork.artist?.targetSupply?.isTargetSupply ??
-    false
-
   const displayText = artwork.consignmentSubmission?.displayText
 
   const isSold = !!displayText && STATUSES[displayText!.toLowerCase()]?.text === "Artwork Sold"
 
-  return isPOneArtist && isSold
+  return isPOneArtist(artwork) && isSold
 }
